+++ src/check/checkTaskArgs.ts
@@ -1,4 +1,4 @@
-// generated by src/check/args.schema.ts
+// generated by src/check/checkTaskArgs.schema.ts
 
 export interface CheckTaskArgs {
 	typecheck?: boolean;
@@ -28,4 +28,4 @@ export interface CheckTaskArgs {
 	'no-lint'?: boolean;
 }
 
-// generated by src/check/args.schema.ts
+// generated by src/check/checkTaskArgs.schema.ts
